Renamed directory to match updated component name.
